Did quite a bit of cleaning up.
Box is now much simpler, most of the data sanitization should occur
outside of Box itself. A box can be instantiated by passing in
[a, b, c] and [alpha, beta, gamma].

This will make testing much easier and a lot less clutter in the actual Box Class.

The tests need to be updated to show this.

Still WIP.