Correct energy unit conversion for epsilon
Currently, when trying to write out to a unit system in lammps, the
quantity is converted from `kcal/mol` to `kcal` which essentially sets
the epsilon value to 0.0.

This change converts the `epsilon` unit to `kcal/mol`, although it is
noted that this might have to change in the future. Since the unit
systems of lammps are varied and have different units based on the
system the engine is operating in.