Update README with additional information
Previously, the README for `GMSO` has been empty.

Although the documentation and docstrings are quite fleshed out, the
landing page on github as well as the landing page at mosdef.org/gmso
were quite blank.

This combines some information from the documentation as well as a few
quick start guides to help direct the user.