Pin pydantic to versions less than 1.9
With the release of `pydantic` 1.9.X, there have been some nuanced
changes to how json handling is done.

This has led to breaking changes in the serialization of some of our
data structures.
This does not happen with pydantic 1.8.2 or less.

The current patch is to pin to versions less than 1.9 until we have a
fix.